split out async resolver code
