+++ src/future.rs
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-use crate::connmgr::resolver;
 use crate::connmgr::tls::{TlsStream, TlsStreamError, VerifyMode};
 use crate::core::arena;
 use crate::core::event::{self, ReadinessExt};
@@ -458,28 +457,6 @@ pub fn io_split<T: AsyncRead + AsyncWrite>(handle: &RefCell<T>) -> (ReadHalf<T>,
 #[track_caller]
 pub fn get_reactor() -> Reactor {
     Reactor::current().expect("no reactor in thread")
-}
-
-pub struct AsyncResolver<'a> {
-    resolver: &'a resolver::Resolver,
-}
-
-impl<'a> AsyncResolver<'a> {
-    pub fn new(resolver: &'a resolver::Resolver) -> Self {
-        Self { resolver }
-    }
-
-    pub fn resolve(&self, host: &str) -> QueryFuture {
-        let query = match self.resolver.resolve(host) {
-            Ok(q) => Some(q),
-            Err(()) => None,
-        };
-
-        QueryFuture {
-            evented: None,
-            query,
-        }
-    }
 }
 
 pub struct AsyncTcpListener {
@@ -1076,75 +1053,6 @@ impl<'a> EventWaiter<'a> {
 
     pub fn wait(&'a self, interest: mio::Interest) -> WaitFuture<'a> {
         WaitFuture { w: self, interest }
-    }
-}
-
-pub struct QueryFuture {
-    evented: Option<CustomEvented>,
-    query: Option<resolver::Query>,
-}
-
-impl Future for QueryFuture {
-    type Output = Result<resolver::Addrs, io::Error>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-        let f = &mut *self;
-
-        let query = match &f.query {
-            Some(query) => query,
-            None => return Poll::Ready(Err(io::Error::from(io::ErrorKind::OutOfMemory))),
-        };
-
-        let evented = match &f.evented {
-            Some(evented) => evented,
-            None => {
-                let evented = CustomEvented::new(
-                    query.get_read_registration(),
-                    mio::Interest::READABLE,
-                    &get_reactor(),
-                )
-                .unwrap();
-
-                evented.registration().set_ready(true);
-
-                f.evented = Some(evented);
-
-                f.evented.as_ref().unwrap()
-            }
-        };
-
-        evented
-            .registration()
-            .set_waker(cx.waker(), mio::Interest::READABLE);
-
-        if !evented.registration().is_ready() {
-            return Poll::Pending;
-        }
-
-        match query.process() {
-            Some(ret) => Poll::Ready(ret),
-            None => {
-                evented.registration().set_ready(false);
-
-                Poll::Pending
-            }
-        }
-    }
-}
-
-impl Drop for QueryFuture {
-    fn drop(&mut self) {
-        if let Some(evented) = &self.evented {
-            let query = self.query.as_ref().unwrap();
-
-            // normally, a registration will deregister itself when dropped.
-            // however, the query's registration is not dropped when the
-            // query is dropped, so we need to explicitly deregister
-            evented
-                .registration()
-                .deregister_custom(query.get_read_registration())
-                .unwrap();
-        }
     }
 }
 
@@ -2382,7 +2290,6 @@ mod tests {
     use crate::core::zmq::SpecInfo;
     use std::cmp;
     use std::fs;
-    use std::net::IpAddr;
     use std::rc::Rc;
     use std::str;
     use std::task::Context;
@@ -2510,30 +2417,6 @@ mod tests {
             .unwrap();
 
         executor.run(|_| Ok(())).unwrap();
-    }
-
-    #[test]
-    fn test_resolve() {
-        let reactor = Reactor::new(1);
-        let executor = Executor::new(1);
-
-        executor
-            .spawn(async {
-                let resolver = resolver::Resolver::new(1, 1);
-                let resolver = AsyncResolver::new(&resolver);
-
-                let f1 = resolver.resolve("127.0.0.1");
-                let f2 = resolver.resolve("127.0.0.1"); // will error, since queries_max=1
-
-                let addrs = f1.await.unwrap();
-                let e = f2.await.unwrap_err();
-
-                assert_eq!(addrs.as_slice(), &[IpAddr::from([127, 0, 0, 1])]);
-                assert_eq!(e.kind(), io::ErrorKind::OutOfMemory);
-            })
-            .unwrap();
-
-        executor.run(|timeout| reactor.poll(timeout)).unwrap();
     }
 
     #[test]
